requirements.txt
Flask==2.3.0
Flask-Login==0.6.2
Flask-WTF==1.1.1
mongoengine==0.27.0
pymongo==4.5.0
python-dotenv==1.0.0
werkzeug==2.3.0
gunicorn==21.2.0
Werkzeug==2.3.0
click==8.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
